% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoK.R
\name{autoK}
\alias{autoK}
\title{Automatically cluster 1-dimensional continuous data}
\usage{
autoK(x)
}
\arguments{
\item{x}{Vector to be clustered. Must contain at least 1 non-missing value.}
}
\value{
An integer vector identifying the cluster corresponding to each element in \code{x}.
}
\description{
Automatically cluster 1-dimensional continuous data
}
\examples{
1
}
