% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{comms}
\alias{comms}
\title{Simulated records of radio station communications.}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 50000 rows and 4 columns.
}
\usage{
comms
}
\description{
Dataset intended to demonstrate usage of \code{sift::conjecture}.
}
\keyword{datasets}
