% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conjecture.R
\name{conjecture}
\alias{conjecture}
\title{conjecture}
\usage{
conjecture(data, sort_by, names_from, names_first)
}
\arguments{
\item{data}{A data frame to reshape.}

\item{sort_by}{Column name, as symbol, to serves as sorting dimension.}

\item{names_from}{Column name, as symbol. Used to differentiate anterior/posterior observations. Columnn must only contain 2 levels (missing values not allowed).}

\item{names_first}{level in variable specified by \code{names_from} indicating anterior observation.}
}
\description{
reshaping / obscure / idiosyncratic
}
\examples{
conjecture(mtcars)
}
