% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nyt2020}
\alias{nyt2020}
\title{2020 New York Times Headlines}
\format{
A data frame with 1,830 rows and 6 variables:
\describe{
  \item{headline}{Article Headline}
  \item{abstract}{Brief summary of article}
  \item{byline}{Contributing Writers}
  \item{pub_date}{Date of Publication}
  \item{section_name}{NYT section in which article was published}
  \item{web_url}{Article URL}
  ...
}
}
\source{
Obtained using \href{https://developer.nytimes.com/}{NYT Developer Portal} (Archive API)
}
\usage{
nyt2020
}
\description{
Includes selected headlines and additional metadata for NYT articles throughout 2020. This dataset is not a comprehensive account of all major events from 2020.
}
\keyword{datasets}
