% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{us_uk_pop}
\alias{us_uk_pop}
\alias{us_uk_leaders}
\title{Fragments of US & UK population & leaders}
\source{
See \code{tidyr::\link[tidyr]{who}} and \code{ggplot2::\link[ggplot2]{presidential}}.
}
\usage{
us_uk_pop

us_uk_leaders
}
\description{
These datasets are intended to demonstrate usage of \code{sift::break_join}.
}
\keyword{datasets}
