% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sift.R
\name{sift}
\alias{sift}
\title{Augmented data frame subsetting. Imagine `dplyr::filter()` that includes neighboring observations.}
\usage{
sift(.data, sift.col, scope, ...)
}
\value{
A data frame.
}
\description{
Augmented data frame subsetting. Imagine `dplyr::filter()` that includes neighboring observations.
}
\examples{
# Grouped dataframes
# Multiple scopes
# .cluster/.anchor column
}
